requirements.txt
langchain
langchain_google_genai
langchain_experimental
langchain_community
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
mysql-connector-python
pymysql
sentence-transformers
chromadb
